feat: A tail recursive function was created that counts numbers
The counter_number file is a recursive function created to count how many digits a phone number has. It also has the ability to count digits of any type, whether text, number or symbols. The program is experimental, it is only an initial attempt at a challenge which will be loaded in its initial version and then be modified and improved using new and different branches to avoid conflicts with said file.